+++ Cargo.toml
@@ -4,7 +4,7 @@ members = [
     "ft-sys",
     "ft-sys-shared",
 ]
-exclude = ["sample-wasm", "fo", ]
+exclude = ["sample-wasm", "foo", ]
 resolver = "2"
 
 [workspace.package]
