+++ foo/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "fo"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+chrono = "0.4"
+
+[dependencies.diesel]
+version = "2"
+features = [
+    "sqlite",
+    "chrono",
+    "postgres",
+    "returning_clauses_for_sqlite_3_35" # this feature runs get_result statement
+]
+
