+++ foo/src/main.rs
@@ -1,0 +1,56 @@
+use diesel::prelude::*;
+
+fn main() {
+    println!("Hello, world!");
+}
+
+diesel::table! {
+    table_with_ts (ts) {
+        // ts -> diesel::sql_types::TimestamptzSqlite,
+        ts -> diesel::sql_types::Timestamptz,
+        // ts -> diesel::sqlite::sql_types::Timestamptz,
+    }
+}
+
+diesel::table! {
+    ft_user (updated_at) {
+        id -> Int8,
+        username -> Text,
+        updated_at -> Int8,
+    }
+}
+
+// pub fn i(c: &mut diesel::sqlite::SqliteConnection) {
+pub fn i(c: &mut PgConnection) {
+    // let data: Vec<(i64, String /* chrono::DateTime<chrono::Utc> */)> = ft_user::table
+    let data: Vec<chrono::DateTime<chrono::Utc>> = table_with_ts::table
+        // .select((ft_user::id, ft_user::username, ft_user::updated_at))
+        .select(table_with_ts::ts)
+        .get_results(c)
+        .unwrap();
+}
+
+#[derive(diesel::Insertable, diesel::Queryable, diesel::Selectable, Debug)]
+#[diesel(table_name = ft_user)]
+#[diesel(treat_none_as_default_value = false)]
+pub struct User2 {
+    /// id is guaranteed to be the same as `fastn_user(id)`
+    pub id: i64,
+    pub username: String,
+    pub updated_at: i64,
+}
+pub fn insertable(c: &mut diesel::sqlite::SqliteConnection) {
+    let user = User2 {
+        id: 1,
+        username: "yo".to_string(),
+        updated_at: 1,
+    };
+
+    let c = diesel::insert_into(ft_user::table)
+        .values(user)
+        .returning(ft_user::id)
+        .get_result::<i64>(c)
+        .unwrap();
+}
+
+
